Fixing the bugs in ze codez
1. database disconnect after every query
2. adding session check on add/edit/delete pages
3. checking if user has logged in on the start page of cms

+++ admin/post-delete.php
@@ -1,13 +1,30 @@
 <?php
 
+session_start();
+
 include '../DatabaseConnection.php';
+
+if (!isset($_SESSION['logged_in'])) {
+
+	header('Location:index.php');
+	exit();
+
+}
+
 $id = $_GET['id'];
+
 $dbobj = new DBConnect();
 
 $dbobj -> connect();
 
-$query = sprintf("delete from `data` where `id` = %d",$id);
+$query = sprintf("delete from `data` where `id` = %d", $id);
 
 $results = $dbobj -> sqlQuery($query);
+
+if ($results) {
+
+	$dbobj -> disconnect();
+}
+
 header('Location:index.php');
 ?>
